pyproject.toml
[tool.poetry]
name = "auto-uav"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pylint = "^2.14.5"
mypy = "^0.971"
black = "^22.6.0"
torch = "^1.12.0"
numpy = "^1.23.1"
pandas = "^1.4.3"
matplotlib = "^3.5.2"
opencv-python = "4.5.5.62"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
